refactor(api/hooks): api 별로 디렉토리 분리
