+++ src/api/hooks/products/useGetProductOptions.ts
@@ -1,9 +1,9 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
 
+import type { ProductDetailRequestParams } from '@/api/hooks/products/useGetProductDetail';
 import type { ProductOptionsData } from '@/types';
 
-import { BASE_URL, fetchInstance } from '../instance';
-import type { ProductDetailRequestParams } from './useGetProductDetail';
+import { BASE_URL, fetchInstance } from '../../instance';
 
 type Props = ProductDetailRequestParams;
