+++ src/components/features/Category/CategoryProductsSection/index.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 import { Link } from 'react-router-dom';
 
-import { useGetProducts } from '@/api/hooks/useGetProducts';
+import { useGetProducts } from '@/api/hooks/products/useGetProducts';
 import { DefaultGoodsItems } from '@/components/common/GoodsItem/Default';
 import { Container } from '@/components/common/layouts/Container';
 import { Grid } from '@/components/common/layouts/Grid';
