+++ src/components/features/Goods/Detail/index.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 
-import type { ProductDetailRequestParams } from '@/api/hooks/useGetProductDetail';
+import type { ProductDetailRequestParams } from '@/api/hooks/products/useGetProductDetail';
 import { breakpoints } from '@/styles/variants';
 
 import { GoodsDetailHeader } from './Header';
