+++ src/mocks/browser.ts
@@ -1,6 +1,6 @@
 import { setupWorker } from 'msw';
 
-import { categoriesMockHandler } from '@/api/hooks/categories.mock';
-import { productsMockHandler } from '@/api/hooks/products.mock';
+import { categoriesMockHandler } from '@/api/hooks/categories/categories.mock';
+import { productsMockHandler } from '@/api/hooks/products/products.mock';
 
 export const worker = setupWorker(...categoriesMockHandler, ...productsMockHandler);
